+++ docs/sphinx/conf.py
@@ -12,6 +12,12 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
+# Disable PEP 8 checks
+# flake8: noqa
+
+# Disable Black formatting
+# fmt: off
+
 import sys
 import os
 # import shlex