+++ ftpsync/ftp_target.py
@@ -8,9 +8,7 @@ import ftplib
 import json
 import os
 import time
-from posixpath import join as join_url
-from posixpath import normpath as normpath_url
-from posixpath import relpath as relpath_url
+from posixpath import join as join_url, normpath as normpath_url, relpath as relpath_url
 from tempfile import SpooledTemporaryFile
 
 from ftpsync.compat import CompatConnectionError