+++ test/fixture_tools.py
@@ -67,7 +67,7 @@ class CaptureStdout(list):
 
     def __exit__(self, *args):
         self.extend(self._stringio.getvalue().splitlines())
-        del self._stringio    # free up some memory
+        del self._stringio  # Free up some memory
         sys.stdout = self._stdout
         sys.stderr = self._stderr