+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Fix #26: Crash when not setting verbose option.
 - Print SYST and FEAT when -vv is passed
 - Accept list type options for `exclude` argument in CLI mode
+- Apply and enforce Black formatter
 
 2.0.0 (2018-01-01)
 ------------------